+++ stories/search.js
@@ -124,7 +124,7 @@ export default class App extends Component {
             //
             // Custom comparison for matching during search.
             // This is optional, and defaults to a case sensitive search of
-            // the title and subtitle values.
+            // the title values.
             // see `defaultSearchMethod` in https://github.com/frontend-collective/react-sortable-tree/blob/master/src/utils/default-handlers.js
             searchMethod={customSearchMethod}
             //